Separate wave control flow lowering into dedicated pass
Creates a new LowerWaveControlFlowPass that specifically handles
wave.iterate -> scf.for conversion, avoiding the issue where mixed
lowering was creating unlegalizable IR.

The problem is that if LowerWaveToMLIRPass tries to do both:
1. Control flow lowering (wave.iterate -> scf.for)
2. Operation lowering (wave.read, wave.mma, etc. -> standard ops)

This will cause ` wave.iterate` bodies to be cloned with unconverted wave
operations inside `scf.for` loops, creating IR that couldn't be legalized.

The solution separates the pipeline into two phases:
1. LowerWaveControlFlowPass: wave.iterate -> scf.for (leaves other wave ops unchanged)
2. LowerWaveToMLIRPass: Remaining wave ops -> standard MLIR ops

This enables a clean two-phase lowering pipeline where control flow
is handled first, then remaining wave operations are lowered in the
standard `scf.for` context.
