+++ water/lib/Dialect/Wave/Transforms/LowerWaveToMLIR.cpp
@@ -76,8 +76,8 @@ struct LowerWaveToMLIRPass
         >();
     target.addIllegalOp<wave::AddOp, wave::AllocateOp, wave::CastOp,
                         wave::DivOp, wave::Exp2Op, wave::ExtractSliceOp,
-                        wave::IterateOp, wave::MmaOp, wave::MulOp, wave::ReadOp,
-                        wave::RegisterOp, wave::WriteOp, wave::YieldOp>();
+                        wave::MmaOp, wave::MulOp, wave::ReadOp,
+                        wave::RegisterOp, wave::WriteOp>();
 
     // Mark functions as illegal if they have Wave tensor types in their
     // signature.
